# server/src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shoreline Care OS API",
    "version": "6.2.0",
    "description": "Open-source dietary, clinical nutrition, and food operations platform API for care facilities. Generated from the Express route definitions by server/scripts/generate-openapi.mjs — do not edit by hand; run `npm run docs:generate` in server/."
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local Facility Gateway"
    },
    {
      "url": "https://api.shorelineops.com/api",
      "description": "Enterprise Cloud Sync"
    }
  ],
...
  "paths": {
...
    "/api/ehr/census": {
      "get": {
        "operationId": "get_ehr_census",
        "summary": "GET /api/ehr/census",
        "tags": [
          "ehr"
        ],
        "description": "GET /api/ehr/census\nPulls current active census from connected EHR.\nA07: never serve synthetic data as live EHR data. With no EHR configured\n→ 503 EHR_NOT_CONNECTED. The built-in connector is a synthetic stub, so a\nserved payload is explicitly flagged `demo: true`.\n\nAuthentication: JWT Bearer token; roles: exactly dietitian or admin (strict equality, not rank-based); license tier: enterprise+ (402 otherwise).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "License tier required"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "500": {
            "description": "Server error"
          },
          "503": {
            "description": "Service unavailable — integration not configured"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "dietitian",
            "admin"
          ],
          "tier": "enterprise"
        }
      }
    },
    "/api/ehr/nutrients/analyze": {
      "post": {
        "operationId": "post_ehr_nutrients_analyze",
        "summary": "POST /api/ehr/nutrients/analyze",
        "tags": [
          "ehr"
        ],
        "description": "POST /api/ehr/nutrients/analyze\nComputes therapeutic nutritional analysis for a planned meal & diet order\n\nAuthentication: JWT Bearer token; any authenticated role.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer"
        }
      }
    },
    "/api/ehr/nutrients/usda": {
      "post": {
        "operationId": "post_ehr_nutrients_usda",
        "summary": "POST /api/ehr/nutrients/usda",
        "tags": [
          "ehr"
        ],
        "description": "POST /api/ehr/nutrients/usda\nComprehensive USDA FoodData Central meal breakdown & clinical compliance flags\n\nAuthentication: JWT Bearer token; any authenticated role.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer"
        }
      }
    },
    "/api/ehr/reconciliation-queue": {
      "get": {
        "operationId": "get_ehr_reconciliation_queue",
        "summary": "GET /api/ehr/reconciliation-queue",
        "tags": [
          "ehr"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          }
        ],
        "description": "GET /api/ehr/reconciliation-queue\nLists all pending, approved, and rejected inbound EHR triage items for Registered Dietitians\n\nAuthentication: JWT Bearer token; roles: exactly dietitian or admin (strict equality, not rank-based).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "dietitian",
            "admin"
          ]
        }
      }
    },
    "/api/ehr/reconciliation-queue/{id}/resolve": {
      "post": {
        "operationId": "post_ehr_reconciliation_queue_id_resolve",
        "summary": "POST /api/ehr/reconciliation-queue/:id/resolve",
        "tags": [
          "ehr"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "POST /api/ehr/reconciliation-queue/:id/resolve\nRD resolves an inbound EHR change (APPROVE or REJECT)\nA05: dietitian or admin only — reuses A04's requireDietitianOrAdmin (strict\nrole equality, not rank-based, so frontdesk/manager are refused). Denied\ncalls → 403 + audit-logged by the middleware.\n\nAuthentication: JWT Bearer token; roles: exactly dietitian or admin (strict equality, not rank-based).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "dietitian",
            "admin"
          ]
        }
      }
    },
    "/api/ehr/webhook": {
      "post": {
        "operationId": "post_ehr_webhook",
        "summary": "POST /api/ehr/webhook",
        "tags": [
          "ehr"
        ],
        "description": "POST /api/ehr/webhook\nIngests inbound diet order, texture, or ADT update from PointClickCare / MatrixCare.\nA04: gated on HMAC-SHA256 signature (X-EHR-Signature: sha256=<hex>) against\nEHR_WEBHOOK_SECRET. Missing/invalid signature → 401 + audit-logged.\nUnset secret → 503, route refuses all traffic (fail closed).\n\nAuthentication: HMAC-SHA256 webhook signature (X-EHR-Signature header, EHR_WEBHOOK_SECRET).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized — invalid or missing shared secret / signature"
          }
        },
        "security": [
          {
            "ehrWebhookSignature": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "ehr-webhook-hmac"
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/auth/login. Role claim gates requireRole ranks."
      },
      "ehrWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-EHR-Signature",
        "description": "HMAC-SHA256 of the raw body with EHR_WEBHOOK_SECRET."
      },
      "kioskApiSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer <KIOSK_API_SECRET> for kiosk punch webhooks."
      },
      "stripeWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature with STRIPE_WEBHOOK_SECRET."
      },
      "setupSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-setup-secret",
        "description": "SETUP_BOOTSTRAP_SECRET for first-time facility initialization."
      }
    },
...
  },
...
}